WIP: Integrate PAYONE SEPA in common UI flow
